+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+- Configurable connection pool settings: `pool_size` and `pool_idle_timeout`
+
 ## [0.1.0] - 2022-08-03
 
 - Initial release